+++ tests/_test_process_executor.py
@@ -34,7 +34,7 @@ from collections import defaultdict
 
 from loky._base import (PENDING, RUNNING, CANCELLED, CANCELLED_AND_NOTIFIED,
                         FINISHED, Future)
-from loky.process_executor import ShutdownExecutorError
+from loky.process_executor import ShutdownExecutorError, TerminatedWorkerError
 from loky.process_executor import BrokenProcessPool, LokyRecursionError
 from .test_reusable_executor import ErrorAtPickle
 from .test_reusable_executor import ExitAtPickle
@@ -269,7 +269,7 @@ class ExecutorShutdownTest:
 
         # The crashing job should be executed after the non-failing jobs
         # have completed. The crash should be detected.
-        with pytest.raises(BrokenProcessPool):
+        with pytest.raises(TerminatedWorkerError):
             crash_result.result()
 
         _executor_mixin._test_event.clear()
@@ -564,10 +564,10 @@ class ExecutorTest:
         # Get one of the processes, and terminate (kill) it
         p = next(iter(self.executor._processes.values()))
         p.terminate()
-        with pytest.raises(BrokenProcessPool):
+        with pytest.raises(TerminatedWorkerError):
             future.result()
         # Submitting other jobs fails as well.
-        with pytest.raises(BrokenProcessPool):
+        with pytest.raises(TerminatedWorkerError):
             self.executor.submit(pow, 2, 8)
 
     def test_map_chunksize(self):