+++ tests/test_worker_timeout.py
@@ -1,3 +1,4 @@
+import sys
 import time
 import pytest
 import threading
@@ -101,10 +102,9 @@ class TestTimeoutExecutor():
                 assert len(results) == n_tasks
 
     def test_worker_timeout_shutdown_deadlock(self):
-        """Check that worker timeout don't cause deadlock even when shutting down.
+        """Check that worker timeout don't cause deadlock when shutting down.
         """
-        with pytest.warns(UserWarning,
-                          match=r'^A worker stopped while some jobs'):
+        with pytest.warns(UserWarning) as record:
             with get_reusable_executor(max_workers=2, timeout=.001) as e:
                 # First make sure the executor is started
                 e.submit(id, 42).result()
@@ -113,3 +113,9 @@ class TestTimeoutExecutor():
                 # we are sure that the worker timeout.
                 f = e.submit(id, SlowlyPickling(1))
         f.result()
+
+        # The warning detection is unreliable on pypy
+        if not hasattr(sys, "pypy_version_info"):
+            assert len(record) > 0, "No warnings was emitted."
+            msg = record[0].message.args[0]
+            assert 'A worker stopped while some jobs' in msg